+++ scripts/ecom_test_env.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-export ECOM_DSN="host=localhost port=5432 user=postgres password=postgres dbname=ecom_dev sslmode=disable"
-export ECOM_CREDENTIALS_FILE=/home/andy/ecom-test-fa3e406ce4fe.json